pyproject.toml
[project]
name = "nl2sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.54.0",
    "python-dotenv>=1.0.1",
    "markitdown[all]>=0.0.2",
    "sqlglot>=28.0.0",
    "tabulate>=0.9.0",
    "tiktoken>=0.12.0",
    "ipykernel>=7.1.0",
]
